flask-api/pyproject.toml
[project]
name = "sms-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "twilio>=9.9.0",
    "uvicorn>=0.40.0",
]
